refactor: стандартизация типов ID до bigint в record-моделях
Все первичные и внешние ключи приведены к int64 для единообразия.
Переименовано поле NAME_RUS → NameRus согласно Go-конвенции.

+++ internal/repository/record/country.go
@@ -1,7 +1,7 @@
 package record
 
 type Country struct {
-	ID          int32  `db:"id"`
+	ID          int64  `db:"id"`
 	Name        string `db:"name"`
 	Description string `db:"description"`
 }
